backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pymupdf>=1.23.4
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank_bm25>=0.2.2
ollama>=0.1.7
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy<2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
